# openapi/ai-capabilities.yaml
openapi: "3.0.3"
info:
  title: AI 能力 API
  description: |
    AI 能力模块，包含：
    - AI 核心（对话、版本、行为、同步）
    - AI 健康检查
    - AI 监控
    - 插件市场
    - RAG 知识库
    - LLM Provider
    - AI 协作
    - 提示词模板
    - Agent Workflow
    - 高级功能
    - Provider 管理
    - AI 模型
    - 文档解析
  version: "1.0.0"
servers:
  - url: http://localhost:3001
    description: 开发环境
...
paths:
...
  /api/ai/prompts/{id}/test:
    post:
      tags: [ai-prompts]
      summary: 测试提示词
      operationId: testPrompt
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                variables:
                  type: object
                  description: "模板变量值"
      responses:
        "200":
          description: 测试成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      rendered:
                        type: string
                        description: "渲染后的提示词"
                      result:
                        type: string
                        description: "AI 响应"
...
  /api/providers:
    get:
      tags: [ai-providers]
      summary: Provider 列表
      operationId: listProviders
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        type:
                          type: string
                        status:
                          type: string
    post:
      tags: [ai-providers]
      summary: 创建 Provider
      operationId: createProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: ["openai", "anthropic", "custom"]
                baseUrl:
                  type: string
                  format: uri
                apiKey:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/providers/{id}:
    get:
      tags: [ai-providers]
      summary: Provider 详情
      operationId: getProvider
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    put:
      tags: [ai-providers]
      summary: 更新 Provider
      operationId: updateProvider
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                baseUrl:
                  type: string
                  format: uri
                apiKey:
                  type: string
                status:
                  type: string
                  enum: ["active", "disabled"]
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    delete:
      tags: [ai-providers]
      summary: 删除 Provider
      operationId: deleteProvider
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /api/models/{id}/test:
    post:
      tags: [ai-models]
      summary: 测试模型连通性
      operationId: testAiModel
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      responses:
        "200":
          description: 测试成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: ["ok", "error"]
                      message:
                        type: string
                      latency:
                        type: number
                        description: "响应时间（ms）"

  # ==================== 文档解析 ====================
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: object
          properties:
            message:
              type: string
...
  parameters:
...
    ObjectIdParam:
      name: id
      in: path
      required: true
      description: "资源 ID（MongoDB ObjectId）"
      schema:
        type: string
        pattern: "^[0-9a-fA-F]{24}$"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    workflowKeyAuth:
      type: apiKey
      in: header
      name: X-Workflow-Key